packaging: disable tomlkit and manpages on RHEL
tomcli has recently appeared as a potential build dependency for ELN
packages. It is unclear at this point whether tomcli will end up as a
required dependency for RHEL 11, but if it does, it will likely be with
minimal dependencies.  tomli-w is already in RHEL, but tomlkit is not, nor
is scdoc for the manpages.  If it ends up in EPEL instead (as it may
already for 10), then the extra dependencies are not a concern.

Also, some fixes are needed when manpages are disabled.